Added Vertigo for Madness
After the latest classleads, Vertigo is now a valid affliction for
Whispering Madness/Enlighten, and so the files are updated to reflect
this.

As per the Developer readme, vertigo was added to the raw-svo.empty.lua
madness list, and both the herb and focus parts of the vertigo dict were
updated to check for madness.