Refiller Toxicology Addition
Added in the necessary stuff for using the refiller module to do
Toxicology fills. Max batches are different, so I had to change some of
the logic at the start of filling and determining inpot amounts.

I haven't been able to test this myself, unfortunately. Needs someone
with remedies and toxicology (preferably both on the same character) to
give a go.

Removed a reference to the Concoctions skill (it was present during the
forestal skill transition, and never removed).

The caloric ingredient list was different than the others; I changed it
for more consistency; there should be no functional changes.