Add Travis config to generate a release and docs
The travis.yml will not work after commit into the svof repo as each encryption
key is bound to each repo. My preference would be to create a machine account
and record the credentials in a special helpfile in the svof clan with read
permissions for a special clan position which corresponds to the owner group in
the org. I'd be willing to set everything up with that bot.