Add basic shikudo form defenses
This adds Shikudo as a class option, which is a simple solution
to avoid multiple checks throughout the system which associate
the Monk class with Tekura.

Additionally, all Shikudo forms have been added as defenses and tested
in client to ensure that they work with `df`, `vshow defup` and
`vdefs [set]`.