Quick patch
Added in lines to make retired people you honours as unranked instead of hanging up. Fixed a bug with the capture of names when you did qwho and there was no one around who was hidden and it wouldnt display cwho correctly. Also changed the download link for the datamining on honours to a link that does provide all the information we were mining for mainly house and class update.