Added ndb.issoldier(name) (#416)
* Added ndb.issoldier(name)

Added in a check function to see if person is a soldier as set in the ndb

* Fixing parse information

When the NDB pulled information from the site it entered it into the database as all lower. Matches for functions in the NDB have it match to titled entries and this fixes that so it saves to the database as titled.