Diagnose fix (#435)
This fixes the looping issue with the dv alias causing diagnose to loop randomly. It was setting the wrong variable to true and the system was not recognizing that we told it to do this and would continue to loop because it thought that we still needed to diag but kept thinking the true diag was an illusion.