Shrugging enhancement
Added a check into the system for executing shrugging to put it on the severside curing queue due to many people using the serverside queueing it would never be used if they were chasing balance as the queueing would be able to fire before svo is able to do so. This allows it to be used as the curing queue will fire before the serverside command queue does.