Fix for issue #617: Autorewield spam (#618)
* Corrected issue report #617

Added a line to the rewield check with the proper check. Rewield spam should be corrected.

* Actual fix for issue #617.

Okay, so. Learning curve. This is the actual fix for issue #617. Added a check for the proper text on wielding an item, and it stopped the rewield spam.

* Correcting to standard.

Woops! I added a few changes because my class is Magi that I didn't intend to add to the overall change.