Minor convenience updates (#673)
* Update svo (install me in module manager).xml

Adds additional possibilities for sipping a salve. Kneel line is, in fact, humblkneel emote, updated to reflect this.

* Update svo (aliases, triggers).xml

Added regex for manually lying down as well. Also allows for advanced action emoting.

* Update svo (aliases, triggers).xml

ftps is non-standard and was, I believe, a typo. Can't imagine this being needed, and I'm not even sure the regex would allow for it!