Class detection changes
Changes elemental lord format in svo.classes. Changed "Water Elemental" to "Water", etc

Also added some triggers to help detect elemental lords and Psion.

Should allow class flagging to work via triggers and TN/TF for elemental lords and psion as requested in #693

I changed this according to the suggestions but did not use the 'svo.classchange' suggestion as that refers to personal class and not who you are fighting.

NOTE: One concern is if someone was using something like svo.enableclasses or svo.classes with the old class names. Since there were no triggers to detect those classes, this it may not be a big deal though