Update shrugging failsafe timer.
Seems the timer on shrugging was wrong causing svof to spam add shrugging to the curing queue. This should resolve #755 mostly. If someone is chasing balance with the in game queue and not letting the curing queue fire then there is a chance it'll still try to add to the queue and end up filling it up.

This could be resolved by tracking the curing queue and that is not a huge benefit at the moment.