Add config containerid
Allows you to change the container that you want to fill from the pot. Before it was only 'emptyvial' but now you can specify which vial or tun you might want to fill up. For example, say you want a certain tun to be health or want to fill your artifact vial with health.