Update reflexive pronoun to be fully \w+
I had done this with other reflexive pronouns (himself/herself/faenself).
Changing here for consistency. Fae/Faen will probably be Faenself for the reflexive pronoun, but using \w+ to say "A pronoun word goes here" is the most future proofing based on how it sounds the devs intend to update the game.