Emergency fix for Mudlet's sync bug wiping out files
Due to a bug in Mudlet's module sync, need to disable sync first or it will wipe out the xmls. Putting this bandaid fix so people don't run into it anymore until it is fixed. Also fixed an unrelated problem with the naming of one of the modules in the list.