Fixes / QoL for Undead
NDB's manual honors does a request to the API for the information available there. The Undead's city is called 'underworld' there, so this gets persisted improperly after Honors, and until you correct it with a QWC, at which point the color registration fixes things. A potentially 'better' fix would be to just coalesce towards Underworld and update the QWC naming instead. I believe this discrepancy may also have been causing the issues with ndb.isenemy not working consistently on Undead.

Also adds an ndb.isundead function.