Cleanup of triggers and many misc fixes
Lots of changes, mostly deleting/disabling triggers that are not needed because they are either defs/affs tracked by gmcp so triggers are not neccessary for them. Also while I was at it, I decided to fix some wrong things I have seen through the code in some affs and other things I was looking while cleaning things. Still not done, there is a lot of triggers to be reviewed and disabled or deleted.